% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BatchAggregate.R
\name{BatchAggregate}
\alias{BatchAggregate}
\title{Batch aggregate raster objects}
\usage{
BatchAggregate(inRaster, scaleVector, outNames, NoCores)
}
\arguments{
\item{inRaster}{A raster object to be resampled}

\item{scaleVector}{A vector of the different resampling factors to be applied}

\item{outNames}{a string of folder location and stack names used for writing outputs, will have the aggregation values appended}

\item{NoCores}{Number of Cores to used to parallel processing}
}
\value{
raster objects written to file following resampling.
}
\description{
Takes a raster object and a vector of aggrgation values calculates resampling a writes to a folder

This function mainly exists as it seems difficult to use parallel cores to resample a raster object. Hence paralle the
different scale functions and use the outputs in seperate functions.
}
\examples{
#resolutionVector  <-2:8
 #outName <- "F:\\\\Projects\\\\shrub_cover_paper\\\\pixelSizeTests\\\\ras_stack"
 #returns:  "F:\\Projects\\shrub_cover_paper\\pixelSizeTests\\ras_stack8.tif"  "F:\\Projects\\shrub_cover_paper\\pixelSizeTests\\ras_stack7.tif" etc
}

