% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch_csv_create.R
\name{csv_batch}
\alias{csv_batch}
\title{Batch Generate CSV Training Files}
\usage{
csv_batch(raster_path, inPred, numSamps, fromVals, toVals)
}
\arguments{
\item{raster_path}{A list of paths to the classified high-resolution imagery (use list.files)}

\item{fromVals}{vector of the values in classified image}

\item{toVals}{vector which values will be changes too}

\item{inPredImage}{Name and path for the input image that will be used for predictions (Landsat/MODIS)}

\item{ndPred}{No data value for the prediction image (normally 0)}
}
\description{
Preforms a batch implmentation of the csv_create function, allowing multiple high-res images to to processed into training files
at the same time.
}
\examples{
Landsat_SA_CRS <-"..//Landsat_SA_CRS//Landsat_SA_AEA.tif"
shrub_mask_list <-list.files(path = "F:\\\\Projects\\\\Shrub_cover\\\\NGI_aerial_imagery\\\\shrub_masks\\\\",pattern = "*tif$",full.names = TRUE)
set.seed(seed = 33)
percent_cover <-csv_batch(raster_path = shrub_mask_list,inPred = Landsat_SA_CRS,numSamps = 5000)
Class numbers that will be mapped using the following scheme:
   0 = no data such as background, clouds and shadow
   1 = class for which percent cover is being calculated
   2 = all other land cover classes
fromVals <- c(0,1, 2, 3)
toVals <-   c(2,1, 2, 2)
}

