\name{fracCover-package}
\alias{fracCover-package}
\alias{fracCover}
\docType{package}
\title{
\packageTitle{fracCover}
}
\description{
\packageDescription{fracCover}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fracCover}
\packageIndices{fracCover}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fracCover}

Maintainer: \packageMaintainer{fracCover}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
