% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/percent_cover.R
\name{percent_cover_model}
\alias{percent_cover_model}
\title{Random Forest Percent Cover Modelelling}
\usage{
percent_cover_model(inImage, pointdata, outImage, LS.no.data, point_CRS)
}
\arguments{
\item{inImage}{A raster stack of Landsat (or other) data}

\item{pointdata}{Data frame of training data from the csv_create function, does not need to be projected, cols 1,2 are x/y 3 is pc}

\item{LS.no.data}{No data value for the Landsat image, normally 0}

\item{point_CRS}{A CRS file for the points data-not needed if the the CRS is equal}

\item{outImage.}{Name and path for the output classified image}
}
\value{
a raster image classified into percentage coverage
}
\description{
This script reads training data from the CSV file created using the "csv_create/batch_csv"
script.  The script then uses the X and Y coordinates from the training data file to select
the pixel values (predictor values) for each sample point in the input image. The predictor
values and the percent cover data from the training data file (response variable) are
combined and used as input to the random forests model. After the model is created percent
cover predictions are made on the input image to create an output image with percent cover
values ranging from 0 to 1.
}

