% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/percent_cover_parallel_fortest.R
\name{percent_cover_parallel_fortest}
\alias{percent_cover_parallel_fortest}
\title{Percentage Cover Calculations-Full Function
This function is identical to "percent_cover_paralell" but omits the write to raster command to make easier integration in loops}
\usage{
percent_cover_parallel_fortest(no_cores, inImage, pointdata, LS.no.data,
  point_CRS)
}
\description{
Percentage Cover Calculations-Full Function
This function is identical to "percent_cover_paralell" but omits the write to raster command to make easier integration in loops
}

