% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validation_extract.R
\name{validation_extract}
\alias{validation_extract}
\title{Extract Validation Values}
\usage{
validation_extract(validationData, rasterLayers, pointCRS)
}
\arguments{
\item{validationData}{data.frame containing the validation values in format "x","y","pc.cov"}

\item{rasterLayers}{raster object of the variable to extract.}

\item{pointCRS}{CRS object for use with the validationData points.}
}
\value{
a dataframe containing the extracted values, plus actual values and cell numbers. In long format
}
\description{
Takes and input dataframe and containing co-ordinates and actual(validation) values, extracts corresponding
predicted values and return a long format data frame
}

